requirements.txt
python-dotenv
sqlalchemy
flask
flask-sqlalchemy
flask-migrate
flask-login
argon2-cffi
